Refactor game page
